requirements.txt
Click==7.0
dnspython==1.16.0
Flask==1.0.3
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pymongo==3.8.0
Werkzeug==0.15.4